pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-assistant"
version = "0.1.0"
description = "A personalized AI Assistant with long-term memory."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit",
    "langchain",
    "langgraph",
    "langgraph-checkpoint-postgres",
    "langchain-openai",
    "langchain-google-genai",
    "psycopg",
    "psycopg-pool",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
